server/requirements.txt
flask==2.3.3
pikepdf==8.7.1
pillow==10.1.0
gunicorn==23.0.0